requirements.txt
tensorflow
pandas
scikit-learn
flask
python-dotenv
mysql-connector-python
